// src/app/models/store.ts
export class Store {
  id: number;
  name: string;
  countryCode: string;
  email: string;
  mgrFirstName: string;
  mgrLastName: string;
  mgrEmail: string;
  category: number;
  stockBackstore: number;
  stockFrontstore: number;
  stockShoppingWindow: number;
  stockAccuracy: number;
  onFloorAvailability: number;
  stockMeanAge: number;
}